+++ tests/integration/test_service_watcher.py
@@ -6,7 +6,7 @@ import pytest
 from hassette.core.service_watcher import ServiceWatcher
 from hassette.events.hassette import HassetteServiceEvent
 from hassette.resources.base import Service
-from hassette.test_utils.harness import preserve_config
+from hassette.test_utils import preserve_config
 from hassette.types.enums import ResourceStatus