+++ tests/integration/test_states.py
@@ -12,7 +12,7 @@ import pytest
 
 from hassette.models import states
 from hassette.state_manager import DomainStates, StateManager
-from hassette.test_utils.helpers import (
+from hassette.test_utils import (
     make_full_state_change_event,
     make_light_state_dict,
     make_sensor_state_dict,